feat: 폴더 목록 조회 API 구현

+++ data/src/main/java/com/keeply/data/folder/model/FolderResponse.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class FolderResponse(
-    val folderId: Int?,
+    val folderId: Long?,
     val folderName: String?,
     val color: String?
 )
