+++ domain/src/main/java/com/keeply/domain/folder/model/Folder.kt
@@ -1,7 +1,7 @@
 package com.keeply.domain.folder.model
 
 data class Folder(
-    val folderId: Int,
+    val folderId: Long,
     val folderName: String,
     val color: String
 )
